Refactor `Badonde.init` ticket handling
